+++ air-script/tests/pub_inputs/pub_inputs_plonky3.rs
@@ -1,4 +1,4 @@
-use p3_air::{Air, BaseAir, BaseAirWithPublicValues, ExtensionBuilder};
+use p3_air::{Air, BaseAir, BaseAirWithPublicValues, AirBuilder, ExtensionBuilder};
 use p3_matrix::Matrix;
 use p3_field::{Field, PrimeCharacteristicRing};
 use crate::helpers::{AirScriptAir, AirScriptBuilder};
@@ -38,22 +38,26 @@ impl<F: Field, AB: AirScriptBuilder<F = F>> AirScriptAir<F, AB> for PubInputsAir
     }
 
     fn eval(&self, builder: &mut AB) {
-        let main = builder.main();
         let public_values: [_; NUM_PUBLIC_VALUES] = builder.public_values().try_into().expect("Wrong number of public values");
         let periodic_values: [_; NUM_PERIODIC_VALUES] = builder.periodic_evals().try_into().expect("Wrong number of periodic values");
+        let main = builder.main();
         let (main_current, main_next) = (
             main.row_slice(0).unwrap(),
             main.row_slice(1).unwrap(),
         );
-        builder.when_first_row().assert_zero_ext::<_>(main_current[0].clone().into() - public_values[8].into());
-        builder.when_first_row().assert_zero_ext::<_>(main_current[1].clone().into() - public_values[9].into());
-        builder.when_first_row().assert_zero_ext::<_>(main_current[2].clone().into() - public_values[10].into());
-        builder.when_first_row().assert_zero_ext::<_>(main_current[3].clone().into() - public_values[11].into());
-        builder.when_last_row().assert_zero_ext::<_>(main_current[0].clone().into() - public_values[12].into());
-        builder.when_last_row().assert_zero_ext::<_>(main_current[1].clone().into() - public_values[13].into());
-        builder.when_last_row().assert_zero_ext::<_>(main_current[2].clone().into() - public_values[14].into());
-        builder.when_last_row().assert_zero_ext::<_>(main_current[3].clone().into() - public_values[15].into());
-        builder.when_transition().assert_zero_ext::<_>(main_next[0].clone().into() - (main_current[1].clone().into() + main_current[2].clone().into()));
+
+        // Main boundary constraints
+        builder.when_first_row().assert_zero(main_current[0].clone().into() - public_values[8].into());
+        builder.when_first_row().assert_zero(main_current[1].clone().into() - public_values[9].into());
+        builder.when_first_row().assert_zero(main_current[2].clone().into() - public_values[10].into());
+        builder.when_first_row().assert_zero(main_current[3].clone().into() - public_values[11].into());
+        builder.when_last_row().assert_zero(main_current[0].clone().into() - public_values[12].into());
+        builder.when_last_row().assert_zero(main_current[1].clone().into() - public_values[13].into());
+        builder.when_last_row().assert_zero(main_current[2].clone().into() - public_values[14].into());
+        builder.when_last_row().assert_zero(main_current[3].clone().into() - public_values[15].into());
+
+        // Main integrity/transition constraints
+        builder.when_transition().assert_zero(main_next[0].clone().into() - (main_current[1].clone().into() + main_current[2].clone().into()));
     }
 }