+++ codegen/plonky3/src/imports.rs
@@ -6,6 +6,7 @@ pub(super) fn add_imports(scope: &mut Scope) {
     scope.import("p3_air", "Air");
     scope.import("p3_air", "BaseAir");
     scope.import("p3_air", "BaseAirWithPublicValues");
+    scope.import("p3_air", "AirBuilder");
     scope.import("p3_air", "ExtensionBuilder");
     scope.import("p3_matrix", "Matrix");
     scope.import("p3_field", "Field");